pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sprintmind"
version = "0.4.4"
description = "自进化敏捷开发团队框架"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "SprintMind Team"}
]
keywords = ["agile", "sprint", "ai", "automation", "development"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
sprintmind = "sprintmind.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
